refactor fonts and 'StyledCard' component
