+++ pages/_app.js
@@ -1,15 +1,15 @@
 import GlobalStyle from "../styles";
 import Head from "next/head";
-import {Cutive, Inter } from "next/font/google";
+// import {Cutive, Inter } from "next/font/google";
 
-const cutive = Cutive({
-  subsets: ["latin"],
-  weight: "400",
-});
-const inter = Inter({
-  subsets: ["latin"],
-  weight: ["200", "300", "400", "500"],
-});
+// const cutive = Cutive({
+//   subsets: ["latin"],
+//   weight: "400",
+// });
+// const inter = Inter({
+//   subsets: ["latin"],
+//   weight: ["200", "300", "400", "500"],
+// });
 
 export default function App({ Component, pageProps }) {
   return (
@@ -18,12 +18,12 @@ export default function App({ Component, pageProps }) {
       <Head>
         <title>Voc-Trainer</title>
       </Head>
-      <style jsx global>{`
+      {/* <style jsx global>{`
         :root {
           --fontfamily-regular: ${inter.style.fontFamily};
           --fontfamily-special: ${cutive.style.fontFamily};
         }
-      `}</style>
+      `}</style> */}
       <Component {...pageProps} />
     </>
   );
