+++ styles.js
@@ -1,5 +1,14 @@
 import { createGlobalStyle } from "styled-components";
+import { Cutive, Inter } from "next/font/google";
 
+const cutive = Cutive({
+  subsets: ["latin"],
+  weight: "400",
+});
+const inter = Inter({
+  subsets: ["latin"],
+  weight: ["200", "300", "400", "500"],
+});
 
 export const accentColors = [
   "#a9ead8",
@@ -30,8 +39,8 @@ const GlobalStyles = createGlobalStyle`
     --boxshadow-primary: hsl(0 0% 0% / 15%) 0px 2px 12px 0px;
     --boxshadow-secondary: hsl(0 0% 0% / 25%) 0px 7px 6px -5px;
 
-    //--fontfamily-regular: 'Inter', sans-serif;
-    //--fontfamily-special: 'Cutive', monospace;
+    --fontfamily-regular: ${inter.style.fontFamily};
+    --fontfamily-special: ${cutive.style.fontFamily};
     
     font-size: 100%;     
       }
@@ -49,46 +58,6 @@ const GlobalStyles = createGlobalStyle`
     font-family: var(--fontfamily-regular);
   }
 
-/* inter-200 - latin */
-@font-face {
-  font-display: swap; /* Check https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-display for other options. */
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 200;
-  src: url('/fonts/inter-v12-latin-200.woff2') format('woff2'); /* Chrome 36+, Opera 23+, Firefox 39+, Safari 12+, iOS 10+ */
-}
-/* inter-300 - latin */
-@font-face {
-  font-display: swap; /* Check https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-display for other options. */
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 300;
-  src: url('/fonts/inter-v12-latin-300.woff2') format('woff2'); /* Chrome 36+, Opera 23+, Firefox 39+, Safari 12+, iOS 10+ */
-}
-/* inter-regular - latin */
-@font-face {
-  font-display: swap; /* Check https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-display for other options. */
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 400;
-  src: url('/fonts/inter-v12-latin-regular.woff2') format('woff2'); /* Chrome 36+, Opera 23+, Firefox 39+, Safari 12+, iOS 10+ */
-}
-/* inter-500 - latin */
-@font-face {
-  font-display: swap; /* Check https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-display for other options. */
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 500;
-  src: url('/fonts/inter-v12-latin-500.woff2') format('woff2'); /* Chrome 36+, Opera 23+, Firefox 39+, Safari 12+, iOS 10+ */
-}
-/* cutive-regular - latin */
-@font-face {
-  font-display: swap; /* Check https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-display for other options. */
-  font-family: 'Cutive';
-  font-style: normal;
-  font-weight: 400;
-  src: url('/fonts/cutive-v17-latin-regular.woff2') format('woff2'); /* Chrome 36+, Opera 23+, Firefox 39+, Safari 12+, iOS 10+ */
-}
 
 `;
